-fixed some openal stuff -infinity thread somewhere: --slap me thrice and hand me to me mama:alDeleteBuffers was dying a strange death instaed of complaining about direct buffers!! -Critical Sections

+++ src/main/java/neo/Sound/sound.java
@@ -171,7 +171,7 @@ public class sound {
         public idStr name;
         public idStr format;
         public int numChannels;
-        public int numSamplesPerSecond;
+        public long numSamplesPerSecond;
         public int num44kHzSamples;
         public int numBytes;
         public boolean looping;
